fix: reconcile tool bootcheck source authority
